- Solucionado bug el enviar informes de errores. - Actualizada la librería para generar PDFs y eliminadas las fuentes antiguas. - Eliminados los archivos innecesarios de phpmailer. - Solucionado el bug que deficultaba modificar la tarifa de un artículo con pvp igual a 0. - Solucionados los bugs que impedían la actualización del stock de los artículos cuando se modificaba un albarán o cuando se eliminaba desde los TPVs. - Completado el plugin del supermercado solidario. - Ya se pueden eliminar clientes y proveedores. - Ya se pueden eliminar impuestos y series. - Modificadas las hojas de estilos para facilitar cambios futuros.
git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@484 fe615fe8-962f-0410-811c-9ded47c8d053

+++ controller/tpv_recambios.php
@@ -67,7 +67,7 @@ class tpv_recambios extends fs_controller
          if(FS_LCD != '')
          {
             $fpt = new fs_printer(FS_LCD);
-            $fpt->add('TOTAL               ');
+            $fpt->add( chr(12).'TOTAL               ' );
             $fpt->add( substr(sprintf('%20s', number_format($_POST['saldo'], 2, ',', '.').' EUROS'), 0, 20) );
             $fpt->imprimir();
          }
@@ -158,7 +158,7 @@ class tpv_recambios extends fs_controller
    
    public function version()
    {
-      return parent::version().'-22';
+      return parent::version().'-23';
    }
    
    private function new_search()
@@ -465,30 +465,34 @@ class tpv_recambios extends fs_controller
       $alb = $albaran->get_by_codigo($_GET['delete']);
       if($alb)
       {
-         $articulo = new articulo();
-         
-         foreach($alb->get_lineas() as $linea)
+         if($alb->ptefactura)
          {
-            $art0 = $articulo->get($linea->referencia);
-            if($art0)
+            $articulo = new articulo();
+            foreach($alb->get_lineas() as $linea)
             {
-               $art0->sum_stock($alb->codalmacen, $linea->cantidad);
-               $art0->save();
+               $art0 = $articulo->get($linea->referencia);
+               if($art0)
+               {
+                  $art0->sum_stock($alb->codalmacen, $linea->cantidad);
+                  $art0->save();
+               }
             }
-         }
-         
-         if( $alb->delete() )
-         {
-            $this->new_message("Ticket ".$_GET['delete']." borrado correctamente.");
             
-            /// actualizamos la caja
-            $this->caja->dinero_fin -= $alb->total;
-            $this->caja->tickets -= 1;
-            if( !$this->caja->save() )
-               $this->new_error_msg("¡Imposible actualizar la caja!");
+            if( $alb->delete() )
+            {
+               $this->new_message("Ticket ".$_GET['delete']." borrado correctamente.");
+               
+               /// actualizamos la caja
+               $this->caja->dinero_fin -= $alb->total;
+               $this->caja->tickets -= 1;
+               if( !$this->caja->save() )
+                  $this->new_error_msg("¡Imposible actualizar la caja!");
+            }
+            else
+               $this->new_error_msg("¡Imposible borrar el ticket ".$_GET['delete']."!");
          }
          else
-            $this->new_error_msg("¡Imposible borrar el ticket ".$_GET['delete']."!");
+            $this->new_error_msg('No se ha podido borrar este albarán porque ya está facturado.');
       }
       else
          $this->new_error_msg("Ticket no encontrado.");
@@ -497,6 +501,9 @@ class tpv_recambios extends fs_controller
    private function imprimir_ticket($albaran, $num_tickets=1, $cajon=TRUE)
    {
       $fpt = new fs_printer();
+      
+      if($cajon)
+         $fpt->abrir_cajon();
       
       $linea = "\nTicket: " . $albaran->codigo;
       $linea .= " " . $albaran->fecha;
@@ -549,9 +556,6 @@ class tpv_recambios extends fs_controller
          $fpt->imprimir();
          $num_tickets--;
       }
-      
-      if($cajon)
-         $fpt->abrir_cajon();
    }
 }
